改进 FResource/相关的; 改进 UConvert  - 改进 FResource/相关的 	主要是改进逻辑, 并在部分资源下载加入进度管理.  - 改进 UConvert 	改进decryptAESCBC函数的功能.  - 增加 FOther 	主要是编写过程中有些价值但是不常用不好归类的代码.

+++ src/danor/matookit/functions/FPack.java
@@ -123,7 +123,8 @@ public class FPack
 		}
 		else
 		{
-			pakFile.delete();
+			if(pakFile.getName().indexOf("sound") == -1)
+				pakFile.delete();
 			
 			File logFile = new File(pakFile.getParent(), "itr.log");
 			
@@ -148,7 +149,8 @@ public class FPack
 				connect = new UConnect(option);
 				pakFile = connect.result;
 
-				decompressPack(pakFile);
+				if(pakFile.getName().indexOf("sound") == -1)
+					decompressPack(pakFile);
 
 				UUtil.Output(logFile, (""+i).getBytes(), false);
 			}
